utility related classes put to separate package
